+++ CHANGELOG.md
@@ -94,4 +94,16 @@ the following form: YYYY.0M.0D.
 ### Updated
 
 - Changed design to include HockeyStatTrackerKernel, HockeyStatTracker, and components.jar.
+
+
+## [2024.10.31]
+
+### Added
+
+- Designed Secondary abstract class that implements HockeyStatTracker
+
+### Updated
+
+- Changed design to include HockeyStatTrackerKernel, HockeyStatTracker
 ```
+